+ added ./autogen.sh-generated stuff to SVN, for pd-extended auto-builds      - pkg-config autoconf macros e.g. PKG_CHECK_MODULES are missing on darwin   + removed extraneous (non-library) gfsm/ subdirectories doc/, src/programs, tests/   + added a lot of UNUSED attributes to lighten the pd-extended autobuild logs a bit
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@10421 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/moocow/gfsm/gfsm/src/libgfsm/gfsmIO.c
@@ -335,7 +335,7 @@ ssize_t gfsmio_getdelim(gfsmIOHandle *ioh, char **lineptr, size_t *n, int delim)
     int c = -2;
     GString *gs=NULL;
 
-    while ( *n > 0 && i < (*n-1) && (c=gfsmio_getc(ioh)) != GFSMIO_EOF ) {
+    while ( *n > 0 && i < (((ssize_t)(*n))-1) && (c=gfsmio_getc(ioh)) != GFSMIO_EOF ) {
       (*lineptr)[i++] = c;
 #ifdef GFSM_DEBUG_GETDELIM
       fprintf(stderr, "---> getdelim(i=%d) got char %d ~ '%c' to linebuf\n", i, (char)c, c);//--DEBUG
@@ -502,13 +502,13 @@ gboolean gfsmio_eof_zfile(gzFile zf)
  * gzFile: Read Methods
  */
 gboolean gfsmio_read_zfile(gzFile zf, void *buf, size_t nbytes)
-{ return zf ? (gzread(zf,buf,nbytes)==nbytes) : FALSE; }
+{ return zf ? (gzread(zf,buf,nbytes)==(int)nbytes) : FALSE; }
 
 /*--------------------------------------------------------------
  * gzFile: Write Methods
  */
 gboolean gfsmio_write_zfile(gzFile zf, const void *buf, size_t nbytes)
-{ return zf ? (gzwrite(zf, buf, nbytes)==nbytes) : FALSE; }
+{ return zf ? (gzwrite(zf, buf, nbytes)==(int)nbytes) : FALSE; }
 
 #endif /* GFSM_ZLIB_ENABLED */
